AMR: Fixed bug in level statistics

+++ ippl/test/AMR/ippl-amr/PartBunchAmr.h
@@ -79,7 +79,7 @@ public:
         for (int i = 0; i < nLevel; ++i)
             msg << "Level " << i << " has "
                 << globalPartPerLevel[i] << " ("
-                << globalPartPerLevel[i] / this->getTotalNum() * 100.0
+                << globalPartPerLevel[i] / double(this->getTotalNum()) * 100.0
                 << " \%) of the total particles" << endl;
         
     }
